Move page handler to controller
